refactor(nutrition): extract meal mutation helpers
